Monotone-Parent: 17d950147c88487268099669b4cf1b9415528f54
Monotone-Parent: d1ec834bd0a7844efec6efcae72d69b62f0ff453
Monotone-Revision: 73b9fa2c49abed5db7f47651fe0a7dfc24da3330

+++ lib/triangular_cell.cpp
@@ -252,7 +252,7 @@ TriangularCell::vertex_type TriangularCell::getLocalVertex(const std::size_t ind
   std::vector<vertex_type::value_type> components;
 
   BOOST_FOREACH(const exact_vertex_type::value_type& c, getLocalVertexExact(index))
-    components.push_back(c.toFloat().toDouble());
+    components.push_back(c.toDouble());
 
   return vertex_type(components);
 }
